move use of EditBookingViewModel to ProjectData
